Rename pp folder to pg
