+++ playwright/tests/e2e.spec.ts
@@ -4,16 +4,16 @@ import { test } from "./fixtures.ts";
 import type { Candidate } from "./models/candidate.ts";
 import type { ListSubmitter } from "./models/listSubmitter.ts";
 import type { NameAuthorisation } from "./models/nameAuthorisation.ts";
-import { CandidateListsOverviewPage } from "./pages/pp/candidateListsOverviewPage.ts";
-import { CsvImportExportPage } from "./pages/pp/csvImportExportPage.ts";
-import { FinalisePage } from "./pages/pp/finalisePage.ts";
-import { ListDesignationPage } from "./pages/pp/listDesignationPage.ts";
-import { ListSubmittersPage } from "./pages/pp/listSubmittersPage.ts";
-import { ManageCandidateListPage } from "./pages/pp/manageCandidateListPage.ts";
-import { NameAuthorisationPage } from "./pages/pp/nameAuthorisationPage.ts";
-import { OverviewPage } from "./pages/pp/overviewPage.ts";
-import { PoliticalGroupPage } from "./pages/pp/politicalGroupPage.ts";
-import { SubstituteSubmittersPage } from "./pages/pp/substituteSubmittersPage.ts";
+import { CandidateListsOverviewPage } from "./pages/pg/candidateListsOverviewPage.ts";
+import { CsvImportExportPage } from "./pages/pg/csvImportExportPage.ts";
+import { FinalisePage } from "./pages/pg/finalisePage.ts";
+import { ListDesignationPage } from "./pages/pg/listDesignationPage.ts";
+import { ListSubmittersPage } from "./pages/pg/listSubmittersPage.ts";
+import { ManageCandidateListPage } from "./pages/pg/manageCandidateListPage.ts";
+import { NameAuthorisationPage } from "./pages/pg/nameAuthorisationPage.ts";
+import { OverviewPage } from "./pages/pg/overviewPage.ts";
+import { PoliticalGroupPage } from "./pages/pg/politicalGroupPage.ts";
+import { SubstituteSubmittersPage } from "./pages/pg/substituteSubmittersPage.ts";
 
 test.describe("End-to-end", () => {
   // These walk the entire application flow (~20+ navigations, a CSV upload and
