+++ playwright/tests/pages/pg/switchElectionPage.ts
